+++ lib/bridge_game.dart
@@ -3,5 +3,4 @@ library bridge_game;
 export 'src/enums/index.dart';
 export 'src/models/index.dart';
 export 'src/scoring/index.dart';
-export 'src/utils/displayable_interface.dart';
 export 'src/utils/range.dart';