+++ lib/src/utils/displayable_interface.dart
@@ -1,3 +1,0 @@
-abstract class Displayable {
-  String display();
-}